dj-jobizz/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Django==5.1.6
django-allauth==0.61.1
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==1.23.5
packaging==20.9
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
dj-rest-auth
pdfplumber
python-docx
spacy
scikit-learn
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
